Removed sanity check, added date/time
Sanity check mostly can just cause errors and was more of a personal function. Date/time of capture were collected. This will remain commented and will not affect operation as it is in a try/catch statement. If anything goes wrong, date/time started and stopped are stored as NAN values.